# api-spec.yaml
openapi: 3.0.3
info:
  title: PsychPaper API
  version: 0.1.0
servers:
  - url: /api
    description: Local API base path
paths:
...
  /admin/users:
    post:
      tags:
        - Admin Users
      summary: Create a new admin user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreate'
      responses:
        '201':
          description: Admin user created
          headers:
            Location:
              description: URL of the created admin user resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AdminUser:
      type: object
      required:
        - id
        - username
        - created_at
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
          minLength: 3
          maxLength: 50
        created_at:
          type: string
          format: date-time
    AdminUserCreate:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string